CHANGES:  -added register method and endpoint

+++ src/main/java/com/miso/eksamensprojektf23/repositories/AppointmentNoteRepository.java
@@ -3,5 +3,5 @@ package com.miso.eksamensprojektf23.repositories;
 import com.miso.eksamensprojektf23.models.AppointmentNote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppointmentNoteRepository extends JpaRepository<AppointmentNote, Integer> {
+public interface AppointmentNoteRepository extends JpaRepository<AppointmentNote, Long> {
 }
