+++ src/main/java/com/miso/eksamensprojektf23/repositories/CalendarRepository.java
@@ -3,5 +3,5 @@ package com.miso.eksamensprojektf23.repositories;
 import com.miso.eksamensprojektf23.models.Calendar;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CalendarRepository extends JpaRepository<Calendar, Integer> {
+public interface CalendarRepository extends JpaRepository<Calendar, Long> {
 }
