+++ src/main/java/com/miso/eksamensprojektf23/repositories/DateRepository.java
@@ -3,5 +3,5 @@ package com.miso.eksamensprojektf23.repositories;
 import com.miso.eksamensprojektf23.models.Date;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DateRepository extends JpaRepository<Date, Integer> {
+public interface DateRepository extends JpaRepository<Date, Long> {
 }
